UOLOS2-5: Refactored the map code to separate client and server side rendering of the map
